// quick-tools/src/app/core/models/payload-commands/ip.payload.ts
export interface IpAddress {
    ip: string;
}

export interface IpAddressInfo {
    status: string; // "success"
    country: string; // "Vietnam"
    countryCode: string; // "VN"
    region: string; // "SG"
    regionName: string; // "Ho Chi Minh"
    city: string; // "Ho Chi Minh City"
    zip: string; // "700000"
    lat: number; // 10.822
    lon: number; // 106.6257
    timezone: string; // "Asia/Ho_Chi_Minh"
    isp: string; // "VIETELmetro"
    org: string; // ""
    as: string; // "AS7552 Viettel Group"
    query: string; // "115.73.219.185
}